updated find_obst.cpp
